requirements.txt
# Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Supabase
supabase==2.7.4

# Auth
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP client
httpx==0.27.2

# Redis
redis[hiredis]==5.1.0

# Background workers
arq==0.26.1

# OpenAI
openai==1.50.0

# Retry
tenacity==9.0.0

# Date / recurrence
python-dateutil==2.9.0

# Logging
structlog==24.4.0

# Error tracking
sentry-sdk[fastapi]==2.14.0

# Dev / test
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.32.0
ruff==0.6.9
mypy==1.11.2
locust==2.31.7
